Fixed similar problem with rollbacks in interactive transactions
